Wprowadzone poprawki i rozpoczęte przepisywanie cpp na jave
Poprawione cpp:
- Atrybuty Wektor2D oraz Wektor3D zmienione na zapisanie jedynie długości wektora zamiast punktów początkowych i końcowych
- Wprowadzone rzucenie wyjątkiem jeśli długość jednego z wektorów wynosi 0 przy obliczaniu kąta między wektorami
- Wszystkie metody poprawnie uwzględniają teraz wszystkie współrzędne wektorów (zapomniałem dodać z1 przy kopiowaniu metody z wektora2D)
- Przesunięcie punktu o wektor tworzy teraz nowy przesunięty punkt zamiast nadpisywać stary
Java:
- przetłumaczyłem punkt2D, Wektor2D oraz Wektor3D

+++ Lista0/Lista0/punkt2D.cpp
@@ -8,17 +8,17 @@ punkt2D::punkt2D() {
     y1 = 1;
 }
 
-punkt2D::punkt2D(int a, int b) {
+punkt2D::punkt2D(double a, double b) {
     x1 = a;
     y1 = b;
 
 }
 
-int punkt2D::wezX() const {
+double punkt2D::wezX() const {
     return x1;
 }
 
-int punkt2D::wezY() const {
+double punkt2D::wezY() const {
     return y1;
 }
