+++ Lista0/Lista0/punkt2D.h
@@ -7,18 +7,18 @@ using namespace std;
 
 class punkt2D {
 private:
-    int x1;  // szerokosc punktu
+    double x1;  // szerokosc punktu
 
-    int y1;  // wysokosc punktu
+    double y1;  // wysokosc punktu
 
 public:
     punkt2D(); // konstruktor domyslny tworzacy punkt na wspolzednych 0,0
 
-    punkt2D(int a, int b); // Konstruktor tworzacy punkt na plaszczyznie dwuwymiarowej
+    punkt2D(double a, double b); // Konstruktor tworzacy punkt na plaszczyznie dwuwymiarowej
 
-    int wezX() const; //zwraca polozenie punktu 2d na osi x
+    double wezX() const; //zwraca polozenie punktu 2d na osi x
 
-    int wezY() const; //zwraca polozenie punktu 2d na osi y
+    double wezY() const; //zwraca polozenie punktu 2d na osi y
 
     void wyswietl() const; //wypisuje wspolrzedne punktu na ekran
